:truck: Rename(admin/component): 수정 관리 아이템 폴더 구조
