+++ src/app/(main)/admin/_components/manage/slider-manage.tsx
